(fix) Heroku issues in package.json
Honestly, its not you -- its me. Webpack wasn't getting installed because
it was listed as devDependency when it should have been a dependency. So,
it (correctly) was not being installed on heroku.

Additionally, I had a typo in my package.json script used by heroku which
caused a nonzero exit code in the shell. I think heroku might work now...